visualization data moved out to own file
