+++ website/src/app/core/state/state.module.ts
@@ -6,8 +6,8 @@ import { NgxsStoragePluginModule } from '@ngxs/storage-plugin';
 import { NgxsModule } from '@ngxs/store';
 
 import { environment } from '../../../environments/environment';
-import { RouterState } from './router/router.state';
 import { PageState } from './page/page.state';
+import { RouterState } from './router/router.state';
 import { VisualizationsState } from './visualizations/visualizations.state';
