+++ website/src/app/pages/visualization/visualization-data-resolver.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
 import { EMPTY, Observable } from 'rxjs';
-import { filter, pluck, take, tap, timeout } from 'rxjs/operators';
+import { pluck, take } from 'rxjs/operators';
+import { Visualization } from 'src/app/core/state/visualizations/visualizations';
 
-import { Visualization, VisualizationsState } from '../../core/state/visualizations/visualizations.state';
+import { VisualizationsState } from '../../core/state/visualizations/visualizations.state';
 
 
 @Injectable({ providedIn: 'root' })
